UnrealEngine/Engine/Source/Programs/WebTestsServer/requirements.txt
django==4.2.0
djangorestframework==3.14.0
channels==4.0.0
daphne==4.0.0
